requirements.txt
google-adk>=1.25.0
google-genai>=0.3.0
google-cloud-logging>=3.8.0

langchain>=0.2.0
langchain-community>=0.2.0

wikipedia>=1.4.0
python-dotenv>=1.0.0